fix: move Query Post to post-service
